backend/requirements.txt
# Django
Django==4.2.10
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.3.1

# Channels for WebSockets (optional but included)
channels==4.0.0
channels-redis==4.2.0

# Redis cache
redis==5.0.1

# Environment variables
python-dotenv==1.0.1

# Image processing (for avatar uploads)
Pillow==10.2.0

# Testing
pytest==8.0.0
pytest-django==4.8.0

# Code quality
flake8==7.0.0
black==24.1.1

# Utilities
python-dateutil==2.8.2

# Payments
stripe>=7.0.0

# Async task queue
celery[redis]>=5.3.6

# HTTP client (bKash API calls)
requests>=2.31.0

# bKash SNS webhook signature verification
cryptography>=42.0.0
